+++ tests/testthat/test-glmmTMB.R
@@ -266,4 +266,22 @@ if (require("testthat") && require("insight") && require("glmmTMB")) {
     ))
   })
 
+  test_that("find_random_slopes", {
+
+    skip_on_cran()
+
+    m <- glmmTMB(
+      count ~ child + camper + (1 + xb| persons),
+      ziformula = ~ child + livebait + (1 + zg + nofish| ID),
+      dispformula = ~xb,
+      data = fish,
+      family = truncated_poisson()
+    )
+
+    expect_equal(
+      find_random_slopes(m),
+      list(random = "xb", zero_inflated_random = c("zg", "nofish"))
+    )
+  })
+
 }